+++ tests/TimeAgoTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Serhii\Tests;
 
 use Carbon\CarbonImmutable;
+use DateTimeInterface;
 use PHPUnit\Framework\TestCase;
 use Serhii\Ago\Lang;
 use Serhii\Ago\TimeAgo;
@@ -70,6 +71,26 @@ class TimeAgoTest extends TestCase
             [CarbonImmutable::now()->subMonths(5)->timestamp, '5 months ago'],
             [CarbonImmutable::now()->subYears(30)->timestamp, '30 years ago'],
             [CarbonImmutable::now()->subMinutes(3)->timestamp, '3 minutes ago'],
+        ];
+    }
+
+    /**
+     * @dataProvider provider_for_trans_method_returns_correct_result_after_passing_a_DateTime_object
+     * @test
+     */
+    public function trans_method_returns_correct_result_after_passing_a_DateTime_object(DateTimeInterface $timestamp, string $expect): void
+    {
+        $this->assertSame($expect, TimeAgo::trans($timestamp));
+    }
+
+    public function provider_for_trans_method_returns_correct_result_after_passing_a_DateTime_object(): array
+    {
+        return [
+            [CarbonImmutable::now()->subDays(3)->toDateTime(), '3 days ago'],
+            [CarbonImmutable::now()->subWeeks(2)->toDateTimeImmutable(), '2 weeks ago'],
+            [CarbonImmutable::now()->subMonths(4)->toDateTime(), '4 months ago'],
+            [CarbonImmutable::now()->subYears(20)->toDateTimeImmutable(), '20 years ago'],
+            [CarbonImmutable::now()->subMinutes(5)->toDateTime(), '5 minutes ago'],
         ];
     }
 }