+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## v2.2.0 (2022-01-16)
 
 - Added support for timestamps to `TimeAgo::trans()` method and not just strings.
+- Added support for DateTime and DateTimeImmutable to `TimeAgo::trans()` method.
 
 ----